refactor: nest the component PopupWarning into PathfindingAlgoDescription.
