+++ src/App.js
@@ -2,7 +2,6 @@ import { createElement, useRef } from './utils';
 import Navbar from './Navbar';
 import Description from './Description';
 import Canvases from './Canvases';
-import PopupWarning from './PopupWarning';
 import styles from './App.module.css';
 
 const App = ({ store }) => {
@@ -32,7 +31,6 @@ const App = ({ store }) => {
           subscribe={store.subscribe}
         />
       </main>
-      <PopupWarning store={store} />
     </>
   );
 };
