+++ src/store/actionTypes.js
@@ -1,6 +1,5 @@
 export const UPDATE_GRID_SIZE = 'UPDATE_GRID_SIZE';
 export const SELECT_NEW_ALGO = 'SELECT_NEW_ALGO';
-export const TOGGLE_POPUP_WARNING = 'TOGGLE_POPUP_WARNING';
 export const GENERATING_NEW_MAZE = 'GENERATING_NEW_MAZE';
 export const MAZE_GENERATED = 'MAZE_GENERATED';
 export const SEARCHING_SOLUTION = 'SEARCHING_SOLUTION';
